finalizacion del reto, se actualiza y agrega logica y se arreglan dependencias

+++ models/src/main/java/com/meli/constants/Constants.java
@@ -2,7 +2,7 @@ package com.meli.constants;
 
 public class Constants {
     public static final String SATELLITES_NAMES = "kenobi,skywalker,sato";
-    public static final double[] KENOVI_POSITION = {5, 4};
-    public static final double[] SKYWALKER_POSITION = {4, -3};
-    public static final double[] SATO_POSITION = {-4, 13};
+    public static final float[] KENOVI_POSITION = {5, 4};
+    public static final float[] SKYWALKER_POSITION = {4, -3};
+    public static final float[] SATO_POSITION = {-4, 13};
 }
